Extract advisor and proposal contract boundaries
